Fix of the compare function
The previous version of the compare function was not consistent (it did not follow the requirements on the compare function), see https://tc39.github.io/ecma262/#sec-array.prototype.sort for details. The tests of this package are too small to notice that on V8 (the default JavaScript engine of Node.js) but 2 of the tests fail on an alternative JavaScript engine that I tried.